requirements.txt
openai~=0.27.7
flask==2.3.2
flask-cors~=3.0.10
requests~=2.31.0
spacy~=3.5.3
python-dotenv~=1.0.0
numpy~=1.24.3
elasticsearch~=7.14.0